[REF] point_of_sale, pos_*: one model per file
*: pos_mrp, pos_sale, pos_self_order

See commit title.
