+++ addons/point_of_sale/models/__init__.py
@@ -3,10 +3,12 @@
 from . import account_bank_statement
 from . import pos_load_mixin
 from . import account_payment
+from . import account_cash_rounding
 from . import account_journal
 from . import account_tax
 from . import account_tax_group
 from . import account_move
+from . import pos_pack_operation_lot
 from . import pos_bus_mixin
 from . import barcode_rule
 from . import binary
@@ -14,6 +16,7 @@ from . import digest
 from . import pos_category
 from . import pos_config
 from . import pos_order
+from . import pos_order_line
 from . import pos_session
 from . import product_pricelist
 from . import product_attribute
