+++ addons/point_of_sale/models/pos_pack_operation_lot.py
@@ -1,0 +1,21 @@
+from odoo import api, fields, models, _
+
+
+class PosPackOperationLot(models.Model):
+    _name = 'pos.pack.operation.lot'
+    _description = "Specify product lot/serial number in pos order line"
+    _rec_name = "lot_name"
+    _inherit = ['pos.load.mixin']
+
+    pos_order_line_id = fields.Many2one('pos.order.line')
+    order_id = fields.Many2one('pos.order', related="pos_order_line_id.order_id", readonly=False)
+    lot_name = fields.Char('Lot Name')
+    product_id = fields.Many2one('product.product', related='pos_order_line_id.product_id', readonly=False)
+
+    @api.model
+    def _load_pos_data_domain(self, data):
+        return [('pos_order_line_id', 'in', [line['id'] for line in data['pos.order.line']])]
+
+    @api.model
+    def _load_pos_data_fields(self, config_id):
+        return ['lot_name', 'pos_order_line_id']
