+++ addons/pos_mrp/models/pos_order_line.py
@@ -1,0 +1,16 @@
+from odoo import models
+
+
+class PosOrderLine(models.Model):
+    _inherit = "pos.order.line"
+
+    def _get_stock_moves_to_consider(self, stock_moves, product):
+        self.ensure_one()
+        bom = product.env['mrp.bom']._bom_find(product, company_id=stock_moves.company_id.id, bom_type='phantom').get(product)
+        if not bom:
+            return super()._get_stock_moves_to_consider(stock_moves, product)
+        boms, components = bom.explode(product, self.qty)
+        ml_product_to_consider = (product.bom_ids and [comp[0].product_id.id for comp in components]) or [product.id]
+        if len(boms) > 1:
+            return stock_moves.filtered(lambda ml: ml.product_id.id in ml_product_to_consider and ml.bom_line_id)
+        return stock_moves.filtered(lambda ml: ml.product_id.id in ml_product_to_consider and (ml.bom_line_id in bom.bom_line_ids))
